refactor: isolate tests that spoil the transaction
When a transaction fails in postgres the connection is not usable
withotu savepoints or rollback. Since our tests begin a transaction to
rollback final results, and Clojure JDBC assumes subtransactions are
merged into single transaction, the transaction per test will be in a
failed state forever after the first fail. Therefore the failing test
should be the last DB statement in a test. If more than one failing
statement is required it makes sense to split them each to their own test.