fix browser test by setting a larger window size
after sorting the catalogue, the element that was previously near
the top of the page is now near the bottom and is therefore out of
reach with the previously used window size. increase window size
to the smallest multiple of 10000 pixels that makes the test pass.

unfortunately, the time for running the browser tests increases
from 2m36s to 3m14s.