fix: two options selected, previous and new
There is a mismatch between `item-selected?` that uses db value and
`react-select` component who wants to use `isOptionSelected` by
providing the new value as parameter.

`item-selected?` is currently implemented so that it compares the db
atom value to each option to determine if they match. However
`react-select` calls the function at a time when the db is not yet
set to the new selected value. As a consequence there will be two
values that are shown as selected, the previous (from db) and the
new (from `react-select`).

The default implementation of `isOptionSelected` works for us so we
don't have to pass a function there.