fix: unauthorized /catalogue if the user has not yet logged in
Previously if you go to /catalogue you will get a forbidden if the user
has not logged in. We want to give unauthorized here instead so the user
will be directed to log in. We give forbidden only if the user
actually doesn't have access rights to the catalogue and we know who
they are.