fix: actions with empty comment
Before #2425 the action comment defaulted to "", which was sent to the
backend if no comment was entered. After, it defaulted to nil, which
violated the API schema and resulted in a HTTP 400. Return to the old
behaviour, but add TODOs for cases where it might make sense to
require a nonempty comment.