migrations: don't recreate unique constraint on resource
... in down migrations.

This is so that we can run a `lein reset` even after adding multiple
resources with the same resid.

This is fine becuase a) we don't really support or need to rollback to
versions from 2019 and b) the code for those versions actually works
fine even without the unique constraint in the db.