feat: load codes from GitHub
Includes an improved DB/CSV setup. The idea is to load the codes
manually from the GitHub and not automatically, since someone has to
check that there are no new code types to handle.

The translations should be managed in the database as the translation
files are meant for more static text that can be statically
checked (i.e. the translation key must be found in the code and that's
not the case with the DUO codes).