feat: debounce tree filtering
Since we allow displaying any number of rows in the tree, merely filtering for
a single character (e.g. "a") will show a lot of rows. Likely the user
wants to type a longer word so let's debounce for a while so that they
can type further. Sending the whole tree to React is slow.