feat: parse mondo codes into dag
Update the Mondo code parsing logic to include Mondo code parents.
Parents can be used to create a directed acylic graph which can
be traversed to find a parent hierarchy for given Mondo code. This
is used to deduce if a given disease in DS DUO code is allowed, ie.
it is the same Mondo code, or a parent of it.