fix: CSS live reload
As with Figwheel, we render the Garden into CSS under target/resources
for the devtools to pick up a change. This was updated to Shadow which
handles this automatically with `:dev-http`, but since we are not
using it, we can use `:watch-dir`.

There is no need for the specific `:render-css-file?` config option
and we can just use the `:dev` option.