feat: introduce action concept
Instead of having everything be a button or a link component, the shared common
data is extracted and called an action. Atomic components can then
form buttons or links when given an action.

This enables us to pass an action but the receiver to decide how to
render it.

We refactor edit buttons as actions. And next continue to introduce dropdowns.