fix: default `:filter-value`
Use the documented default of `:value` (as lower-case string) instead of trying to use
`:display-value` as the default for `:filter-value`.

As a previous comment moved more tables to use `:display-value` this
wrong default became a problem. E.g. catalogue table search, that would like to
use `:filter-value` ended up with all filter values being "" because
the `:display-value` was a component, not a string.