refactor: write extended-logging as macro with only params
This enabled us to log the calling API namespace and not the utility
namespace.

We don't need to log the `:uri` here necessarily as it is often the
previous log line (standard logging).

Remove text "extended-logging" so this functions better together with
the standard logging (and does not repeat anything).