feat: relax catalogue item creation API
Although it was possible to leave form away, the key `:form` was
required (with `nil`) value. Now it's possible to leave the key
`:form` away too.

This enables a better developer experience when looking at the Swagger
API. It would show the required asterisk `*` because the field was
required although the value was optional.