feat: cached events with faster manual coerce
Several coerce options, which are lightly tested:
- Schema coerce (the old and slow one)
- manual coerce (with basic Clojure & Medley facilities)
- Specter coerce (fastest with multi-transform)

Cache low level coerced events in memory. In the event sourcing
approach they should be immutable. If the DB is touched outside of
REMS, the server should be restarted (the standard procedure for other
things as well). All the CRUD functions will keep the cache up to date
so that we can safely serve reads from a warm cache.