feat: incremental all-applications-cache
By rewriting the all-applications-cache update to be incremental, we
get a good performance boost. The calculation of new apps-by-user was
very slow as it recalculated all applications. I.e., it would get
slower with a growing application count.

Incremental calculation requires that we track which users are
potentially affected by these changes. I.e., who had rights in the
old application and who has rights in the new application. Some have
roles removed, some added.

The naming and phasing is difficult here. This should be further
improved in the cache rewrite later. I've tried to use the same names.