fix: use enriched apps for roles
Raw apps have only those user roles that come from events. Handler
role on the other hand comes from an injection. We should therefore
use enriched apps when determining user roles, and the update to
"users by role" did not use it. Fortunately, very few places use this
code, and none of them looked at the enriched roles.

Here is a failing test and a fix for it.