fix: user role after deletion
There was a problem in calculating a user's roles, because only the
just updated applications and their roles were considered. The code
was supposed to calculate the new roles for all updated users based on
all of their applications. This fixes the above and adds a test to
catch it in the future.

It also makes sense to base the users by roles on the result of the
previous logic.

The future cache logic will not be based on this but can take advantage of
the test anyway.