feat: support named parameters in email templates
text-format-map is used as shorthand for typing (juxt ...) everywhere. Vector is also easier to spy than function, useful for tests that needed access to the order of vector parameters.

emails-to-recipients is "common parameters" case, and the rest have differing amounts of parameters. At some point vector parameters may be deprecated altogether.