feat: remove localization format parameters checks from start-up
Only vector parameters were checked, and they had to be exactly the same for no good reason. With named parameters this check is no longer relevant, and it probably was mostly thought to be useful for email templates with many vector parameters.

A better solution would check that localizations do not attempt to use parameters they may never receive.