feat: remove EGA implementation (and user secrets)
EGA implementation is not needed anymore.

User secrets isn't being used for anything else, and it's quite
unlikely that it is needed, so let's remove that also.

This leaves the entitlement push (v2) rather empty, but we can keep it
for now as the way to refer to plugins that should be preferred.

See #3239