fix: use raw sub for table rows with implicit subscriptions
As part of re-frame update, subscriptions that are created inside subscription should be explicit, and this is achieved by using raw subscription. Localization functions are the most common source of bugs regarding implicit subscriptions, because they are used as both values and components. Since warnings are only logged during runtime, there may be more misbehaving code.

Also added a comment to table row subscriptions that do not use implicit subscriptions, for completeness.