refactor: make the handled pipe lazy
If we are only going to return "first 50" from the API for our UI, we
can keep the pipe lazy so that only 50 rows ever get processed. This
makes the handled applications opening a bit faster.

The `get-handled-applications` is also extracted for convenience of
calling in e.g. a performance test.