test: check application reload log after blacklist
Blacklist service reloads applications with users that were added to, or removed from application. There was a bug in the reload implementation which never found any applications, regardless of the users. API tests did not catch this because they did not check applications nor logs. Checking logs for "Reloading [0-9] applications ..." is cheap at this point, but maybe there could be more detailed check also.