feat: optimize give-roles-to-user
We can use a transient map to improve the low level performance. We
also don't need the single variant and it can be inlined.

This makes editing workflows approximately 20% faster. The perf test
data has 3000 applications with 100 handlers each. So to enrich each
application, we must give the handler role to 100 handlers.