feat: remove cache hit statistics
This was unnecessary part from the beginning, and mainly just interesting while debugging early version of the cache.

There is a subtle "bug" regarding config startup: mount is not started before the namespace is loaded. It's not an easy issue to solve, because caches are defined in many different namespaces and initialized using function call.