fix: create data used in test
The Organization owner is used in a select statement in the
`create-organization` call. In one test it is created, the value is
selected and the result is checked. But in this test the value is
selected, which fails, and the resulting "empty results" row actually
blocks clicking the next button. I'm not sure why this doesn't fail in
master, but there must be some tiny styling change or other. However,
for consistency, we should make sure we actually create the data that
the test expects.